+++ src/gateway/server.ts
@@ -1,6 +1,6 @@
 import http from "node:http";
 import ws from "ws";
-import { createLogger, initDatabase } from "../util";
+import { createLogger, initDatabase, initRabbitMQ } from "../util";
 import { onConnection } from "./socket/connection";
 
 const Log = createLogger("GATEWAY");
@@ -24,6 +24,7 @@ export class GatewayServer {
 			Log.msg(`Listening on port ${port}`);
 		});
 
+		await initRabbitMQ();
 		await initDatabase();
 
 		if (!this.server.listening) this.server.listen(port);
