+++ src/http/server.ts
@@ -5,7 +5,13 @@ import morgan from "morgan";
 import http from "node:http";
 
 import { errorHandler, routes } from ".";
-import { config, createLogStream, createLogger, initDatabase } from "../util";
+import {
+	config,
+	createLogStream,
+	createLogger,
+	initDatabase,
+	initRabbitMQ,
+} from "../util";
 import { isFederationRequest } from "./routes";
 
 const Log = createLogger("API");
@@ -62,6 +68,7 @@ export class APIServer {
 			Log.msg(`Listening on port ${port}`);
 		});
 
+		await initRabbitMQ();
 		await initDatabase();
 
 		if (!this.server.listening) this.server.listen(port);
